pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "my_postgres_util"  # インストール時の名前
version = "0.1.2"
dependencies = [
    "psycopg2-binary",
    "python-dotenv"
]

[tool.setuptools]
py-modules = ["postgres_util"] # フォルダではなく「ファイル」を指定